Adding MAX32680 and MAX32690 documentation, removing CAN FD support for MAX32662 and MAX32690, updating MAX78000 examples to support both EvKits, and adding 12-bit ssec initialization for all chips which support it.

+++ Libraries/PeriphDrivers/Source/RTC/rtc_reva.c
@@ -170,7 +170,7 @@ int MXC_RTC_RevA_Stop (mxc_rtc_reva_regs_t *rtc)
     return E_SUCCESS;
 }
 
-int MXC_RTC_RevA_Init (mxc_rtc_reva_regs_t *rtc, uint32_t sec, uint8_t ssec)
+int MXC_RTC_RevA_Init (mxc_rtc_reva_regs_t *rtc, uint32_t sec, uint32_t ssec)
 {
     if (MXC_RTC_RevA_GetBusyFlag(rtc)) {
         return E_BUSY;
